Added ability to use imageio library to save movies

+++ Python/klampt/vis/backends/vis_qt.py
@@ -17,6 +17,16 @@ else:
     from PyQt4.QtCore import *
 
 
+def first_nonexistent_file(pattern, start_index=0):
+    """Finds the first file in a sequence that doesn't exist."""
+    import os
+    while start_index < 999:
+        file_name = pattern % start_index
+        if not os.path.exists(file_name):
+            return file_name
+        start_index += 1
+    raise RuntimeError("Too many files in sequence")
+
 class MyQThread(QThread):
     def __init__(self,func,*args):
         self.func = func
@@ -546,6 +556,7 @@ class _MyWindow(QMainWindow):
         self.movie_timer.timeout.connect(self.movie_update)
         self.movie_frame = 0
         self.movie_time_last = 0
+        self.movie_writer = None
         self.saving_html = False
         self.html_saver = None
         self.html_start_time = 0
@@ -684,33 +695,65 @@ class _MyWindow(QMainWindow):
             time_source = self.getTimeSource()
             if time_source is not None:
                 self.movie_time_last = time_source()
+            try:
+                import imageio
+                fn = first_nonexistent_file('klampt_record_%04d.mp4')
+                print("Saving movie to",fn)
+                try:
+                    self.movie_writer = imageio.get_writer(fn,fps=30,codec='libx264')
+                except Exception as e:
+                    print("Error creating movie writer, trying ffmpeg method")
+                    self.movie_writer = None
+            except ImportError:
+                pass
         else:
             self.movie_timer.stop()
-            dlg =  QInputDialog(self)                 
-            dlg.setInputMode( QInputDialog.TextInput) 
-            dlg.setLabelText("Command")
-            dlg.setTextValue('ffmpeg -y -i image%04d.png -vcodec libx264 -pix_fmt yuv420p klampt_record.mp4')
-            dlg.resize(600,100)                             
-            ok = dlg.exec_()                                
-            cmd = dlg.textValue()
-            #(cmd,ok) = QInputDialog.getText(self,"Process with ffmpeg?","Command", text='ffmpeg -y -f image2 -i image%04d.png klampt_record.mp4')
-            if ok:
-                import os,glob
-                os.system(str(cmd))
-                print("Removing temporary files")
-                for fn in glob.glob('image*.png'):
-                    os.remove(fn)
+            if self.movie_writer is None:
+                print("imageio library not available, using backup ffmpeg method")
+                dlg =  QInputDialog(self)
+                dlg.setInputMode( QInputDialog.TextInput) 
+                dlg.setLabelText("Command")
+                fn = first_nonexistent_file('klampt_record_%04d.mp4')
+                dlg.setTextValue('ffmpeg -y -i image%04d.png -vcodec libx264 -pix_fmt yuv420p '+fn)
+                dlg.resize(600,100)                             
+                ok = dlg.exec_()                                
+                cmd = dlg.textValue()
+                #(cmd,ok) = QInputDialog.getText(self,"Process with ffmpeg?","Command", text='ffmpeg -y -f image2 -i image%04d.png klampt_record.mp4')
+                if ok:
+                    import os,glob
+                    os.system(str(cmd))
+                    print("Removing temporary files")
+                    for fn in glob.glob('image*.png'):
+                        os.remove(fn)
+            else:
+                print("... done saving movie.")
+                self.movie_writer.close()
+                self.movie_writer = None
     
+    def write_movie_image(self,frame_index):
+        if self.movie_writer is None:
+            #save to image
+            self.glwidget.program.save_screen('image%04d.png'%(frame_index))
+        else:
+            #write to imageio writer
+            image = self.glwidget.program.get_screen('numpy')
+            try:
+                self.movie_writer.append_data(image)
+            except Exception:
+                print("Unable to write image to movie writer, saving to image file instead")
+                self.glwidget.program.save_screen('image%04d.png'%(frame_index))
+                self.movie_writer = None
+
     def movie_update(self):
         time_source = self.getTimeSource()
         if time_source is not None:
             tnow = time_source()
             while tnow >= self.movie_time_last + 1.0/30.0:
-                self.glwidget.program.save_screen('image%04d.png'%(self.movie_frame))
+                self.write_movie_image(self.movie_frame)
                 self.movie_frame += 1
                 self.movie_time_last += 1.0/30.0
         else:
-            self.glwidget.program.save_screen('image%04d.png'%(self.movie_frame))
+            self.write_movie_image(self.movie_frame)
             self.movie_frame += 1
     
     def toggle_html_mode(self):
